backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.2
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3
prometheus-client==0.19.0
structlog==23.2.0 
supabase>=2.0.0
sqlalchemy[asyncio]
alembic